+++ TwoPaneLayout/library/src/main/java/com/microsoft/device/dualscreen/twopanelayout/common/TwoPaneMode.kt
@@ -1,4 +1,4 @@
-package com.microsoft.device.dualscreen.twopanelayout
+package com.microsoft.device.dualscreen.twopanelayout.common
 
 /**
  * TwoPaneMode
