+++ TwoPaneLayout/nav_sample/src/main/java/com/microsoft/device/dualscreen/twopanelayout/MainActivity.kt
@@ -26,6 +26,9 @@ import androidx.compose.ui.unit.dp
 import androidx.navigation.NavHostController
 import androidx.navigation.NavOptionsBuilder
 import androidx.navigation.compose.rememberNavController
+import com.microsoft.device.dualscreen.twopanelayout.common.Screen
+import com.microsoft.device.dualscreen.twopanelayout.common.TwoPaneMode
+import com.microsoft.device.dualscreen.twopanelayout.twopanelayoutnav.Destination
 import com.microsoft.device.dualscreen.twopanelayout.ui.theme.TwoPaneLayoutTheme
 import com.microsoft.device.dualscreen.twopanelayout.ui.theme.blue
 import com.microsoft.device.dualscreen.twopanelayout.ui.theme.green
